# docs/swagger.yaml
basePath: /
definitions:
...
  lesson.AnswerOption:
    properties:
      id:
        type: integer
      is_correct:
        type: boolean
      question_id:
        type: integer
      text:
        type: string
    type: object
  lesson.AnswerOptionInput:
    properties:
      is_correct:
        type: boolean
      text:
        maxLength: 500
        minLength: 1
        type: string
    required:
    - text
    type: object
  lesson.CreateInput:
    properties:
      content:
        type: string
      is_published:
        type: boolean
      media:
        items:
          $ref: '#/definitions/lesson.MediaInput'
        type: array
      slug:
        maxLength: 255
        minLength: 2
        type: string
      sort_order:
        type: integer
      tests:
        items:
          $ref: '#/definitions/lesson.TestLinkInput'
        type: array
      title:
        maxLength: 255
        minLength: 2
        type: string
    required:
    - title
    type: object
  lesson.Media:
    properties:
      caption:
        type: string
      id:
        type: integer
      lesson_id:
        type: integer
      media_type:
        type: string
      sort_order:
        type: integer
      url:
        type: string
    type: object
  lesson.MediaInput:
    properties:
      caption:
        type: string
      media_type:
        enum:
        - image
        - video
        type: string
      sort_order:
        type: integer
      url:
        type: string
    required:
    - media_type
    - url
    type: object
  lesson.Model:
    properties:
      author_id:
        type: integer
      content:
        type: string
      course_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      is_published:
        type: boolean
      media:
        items:
          $ref: '#/definitions/lesson.Media'
        type: array
      slug:
        type: string
      sort_order:
        type: integer
      tests:
        items:
          $ref: '#/definitions/lesson.TestLink'
        type: array
      title:
        type: string
      updated_at:
        type: string
    type: object
  lesson.PublicAnswerOption:
    properties:
      id:
        type: integer
      text:
        type: string
    type: object
  lesson.PublicQuestion:
    properties:
      id:
        type: integer
      options:
        items:
          $ref: '#/definitions/lesson.PublicAnswerOption'
        type: array
      solution:
        type: string
      sort_order:
        type: integer
      text:
        type: string
    type: object
  lesson.Question:
    properties:
      author_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      lesson_id:
        type: integer
      options:
        items:
          $ref: '#/definitions/lesson.AnswerOption'
        type: array
      solution:
        type: string
      sort_order:
        type: integer
      text:
        type: string
      updated_at:
        type: string
    type: object
  lesson.QuestionInput:
    properties:
      options:
        items:
          $ref: '#/definitions/lesson.AnswerOptionInput'
        minItems: 2
        type: array
      solution:
        type: string
      sort_order:
        type: integer
      text:
        minLength: 1
        type: string
    required:
    - options
    - text
    type: object
  lesson.ReorderItem:
    properties:
      id:
        type: integer
      sort_order:
        type: integer
    required:
    - id
    type: object
  lesson.SubmitAnswerInput:
    properties:
      question_id:
        type: integer
      selected_option_id:
        type: integer
    required:
    - question_id
    - selected_option_id
    type: object
  lesson.SubmitInput:
    properties:
      answers:
        items:
          $ref: '#/definitions/lesson.SubmitAnswerInput'
        minItems: 1
        type: array
    required:
    - answers
    type: object
  lesson.SubmitResponse:
    properties:
      grade:
        type: number
      max_score:
        type: integer
      questions:
        items:
          $ref: '#/definitions/lesson.PublicQuestion'
        type: array
      score:
        type: integer
    type: object
  lesson.TestLink:
    properties:
      description:
        type: string
      id:
        type: integer
      lesson_id:
        type: integer
      test_id:
        type: integer
    type: object
  lesson.TestLinkInput:
    properties:
      description:
        type: string
      test_id:
        type: integer
    required:
    - test_id
    type: object
  lesson.UpdateInput:
    properties:
      content:
        type: string
      is_published:
        type: boolean
      media:
        items:
          $ref: '#/definitions/lesson.MediaInput'
        type: array
      slug:
        maxLength: 255
        minLength: 2
        type: string
      sort_order:
        type: integer
      tests:
        items:
          $ref: '#/definitions/lesson.TestLinkInput'
        type: array
      title:
        maxLength: 255
        minLength: 2
        type: string
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: Документация API сервиса it_courses.
  title: API it_courses
  version: "1.0"
paths:
...
  /courses/{courseId}/lessons:
    get:
      description: Возвращает все уроки указанного курса
      parameters:
      - description: ID курса
        in: path
        name: courseId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/lesson.Model'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Список уроков курса
      tags:
      - lessons
    post:
      consumes:
      - application/json
      description: Создаёт новый урок в курсе. Доступно teacher и admin
      parameters:
      - description: ID курса
        in: path
        name: courseId
        required: true
        type: integer
      - description: Данные урока
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/lesson.CreateInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/lesson.Model'
        "400":
          description: Bad Request
          schema:
            type: string
        "409":
          description: Conflict
          schema:
            type: string
      summary: Создать урок
      tags:
      - lessons
  /courses/{courseId}/lessons/{id}:
    delete:
      description: Удаляет урок из курса. Доступно teacher и admin
      parameters:
      - description: ID курса
        in: path
        name: courseId
        required: true
        type: integer
      - description: ID урока
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            type: string
      summary: Удалить урок
      tags:
      - lessons
    get:
      description: Возвращает урок с контентом, медиа и ссылками на тесты
      parameters:
      - description: ID курса
        in: path
        name: courseId
        required: true
        type: integer
      - description: ID урока
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lesson.Model'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Детали урока
      tags:
      - lessons
    put:
      consumes:
      - application/json
      description: Обновляет данные урока. Доступно teacher и admin
      parameters:
      - description: ID курса
        in: path
        name: courseId
        required: true
        type: integer
      - description: ID урока
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/lesson.UpdateInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lesson.Model'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "409":
          description: Conflict
          schema:
            type: string
      summary: Обновить урок
      tags:
      - lessons
  /courses/{courseId}/lessons/reorder:
    patch:
      consumes:
      - application/json
      description: Меняет порядок уроков в курсе. Доступно teacher и admin
      parameters:
      - description: ID курса
        in: path
        name: courseId
        required: true
        type: integer
      - description: Массив ID и sort_order
        in: body
        name: input
        required: true
        schema:
          items:
            $ref: '#/definitions/lesson.ReorderItem'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Изменить порядок уроков
      tags:
      - lessons
...
  /lessons/{id}/questions:
    get:
      description: Возвращает вопросы урока для прохождения (без правильных ответов)
      parameters:
      - description: ID урока
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/lesson.PublicQuestion'
            type: array
      summary: Вопросы урока
      tags:
      - lessons
  /lessons/{id}/submit:
    post:
      consumes:
      - application/json
      description: Принимает ответы на вопросы урока и возвращает результат. Только
        student
      parameters:
      - description: ID урока
        in: path
        name: id
        required: true
        type: integer
      - description: Ответы на вопросы
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/lesson.SubmitInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lesson.SubmitResponse'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Отправить ответы на вопросы урока
      tags:
      - lessons
  /lessons/{lessonId}/questions:
    post:
      consumes:
      - application/json
      description: Создаёт вопрос с вариантами ответов для урока. Доступно teacher
        и admin
      parameters:
      - description: ID урока
        in: path
        name: lessonId
        required: true
        type: integer
      - description: Вопрос и варианты ответов
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/lesson.QuestionInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/lesson.Question'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Создать вопрос урока
      tags:
      - lessons
  /lessons/{lessonId}/questions/{id}:
    delete:
      description: Удаляет вопрос из урока. Доступно teacher и admin
      parameters:
      - description: ID урока
        in: path
        name: lessonId
        required: true
        type: integer
      - description: ID вопроса
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            type: string
      summary: Удалить вопрос урока
      tags:
      - lessons
    put:
      consumes:
      - application/json
      description: Обновляет вопрос и варианты ответов. Доступно teacher и admin
      parameters:
      - description: ID урока
        in: path
        name: lessonId
        required: true
        type: integer
      - description: ID вопроса
        in: path
        name: id
        required: true
        type: integer
      - description: Новые данные вопроса
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/lesson.QuestionInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lesson.Question'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Обновить вопрос урока
      tags:
      - lessons
  /lessons/{lessonId}/tests:
    post:
      consumes:
      - application/json
      description: Добавляет ссылку на тест в урок. Доступно teacher и admin
      parameters:
      - description: ID урока
        in: path
        name: lessonId
        required: true
        type: integer
      - description: ID теста
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/lesson.TestLinkInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/lesson.TestLink'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Привязать тест к уроку
      tags:
      - lessons
  /lessons/{lessonId}/tests/{testId}:
    delete:
      description: Удаляет ссылку на тест из урока. Доступно teacher и admin
      parameters:
      - description: ID урока
        in: path
        name: lessonId
        required: true
        type: integer
      - description: ID теста
        in: path
        name: testId
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            type: string
      summary: Отвязать тест от урока
      tags:
      - lessons
...
swagger: "2.0"